requirements.txt
pyautogui
pytesseract
opencv-python
pillow
numpy
pygetwindow
